Crate robonomics_parachain_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants
A set of constant values used in substrate runtime.

impls
Some configurable implementations as associated type for the substrate runtime.

Structs

AdjustmentVariable

AvailableBlockRatio

BasicDeposit

BlockHashCount

Burn

CandidacyBond

CouncilMaxProposals

CouncilMotionDuration

CreationFee

DepositBase

DepositFactor

DesiredMembers

DesiredRunnersUp

ElectionsPhragmenModuleId

ExistentialDeposit

FieldDeposit

GenesisConfig

IndexDeposit

MaxAdditionalFields

MaxRegistrars

MaxSignatories

MaxSubAccounts

MaximumBlockLength

MaximumBlockWeight
We allow for 4 seconds of compute with a 12 second average block time.

MaximumExtrinsicWeight
Assume 10% of weight for average on_initialize calls

MaximumSchedulerWeight

MinimumMultiplier

MinimumPeriod

ModuleToIndex
Provides an implementation of ModuleToIndex to map a module to its index in the runtime.

Origin

ProposalBond

ProposalBondMinimum

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SessionKeys

SpendPeriod

SubAccountDeposit

TargetBlockFullness

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TipReportDepositPerByte

TransactionByteFee

TransferFee

TreasuryModuleId

Version

VotingBond

Enums

Call

Event

OriginCaller

Constants

VERSION
Standalone runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap
Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address
The address format for describing accounts.

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Context
The type used as a helper for interpreting the sender of transactions.

Council

CouncilConfig

Datalog

Elections

ElectionsConfig

Executive
Executive: handles dispatch to the various modules.

GenericAsset

GenericAssetConfig

Header
Block header type as expected by this runtime.

Identity

Indices

IndicesConfig

Launch

Liability

MessageBroker

Multisig

ParachainInfo

ParachainInfoConfig

ParachainUpgrade

RandomnessCollectiveFlip

Scheduler

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TokenDealer

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility